[Tizen] Add codes for Dali Windows Backend
This reverts commit 8a1ba749cd94c2bc08e71baf02c72a1a3d381c77.

Conflicts:
	dali-toolkit/internal/layouting/linear-layout-impl.cpp

+++ dali-toolkit/internal/layouting/vbox-layout-impl.cpp
@@ -233,7 +233,7 @@ void VboxLayout::OnLayout( bool changed, LayoutLength left, LayoutLength top, La
   auto width = right - left;
 
   // Space available for child
-  auto childSpace = width - padding.start - padding.end;
+  auto childSpace = width - (int)padding.start - (int)padding.end;
   auto count = GetChildCount();
 
   for( unsigned int childIndex = 0; childIndex < count; childIndex++)
@@ -248,10 +248,10 @@ void VboxLayout::OnLayout( bool changed, LayoutLength left, LayoutLength top, La
       auto childMargin = childLayout->GetMargin();
 
       childTop += childMargin.top;
-      childLeft = ( childSpace - childWidth ) / 2 + childMargin.start - childMargin.end;
+      childLeft = ( childSpace - childWidth ) / 2 + (int)childMargin.start - (int)childMargin.end;
 
       childLayout->Layout( childLeft, childTop, childLeft + childWidth, childTop + childHeight );
-      childTop += childHeight + childMargin.bottom + mCellPadding.height;
+      childTop += childHeight + (int)childMargin.bottom + mCellPadding.height;
     }
   }
 }
